Adds new type of searchBar to fit in better with big title
Adds in new type of searchBar that is searchController. Ininitally there was a bug with the keyboard will automaticlly get rid of itself for some reason, I just removed the code that would get rid of the keyboard. Searchview seems to have that functionality built in, so there was no need for that code anyways